+++ tests/devices/unit_tests/test_undulator.py
@@ -1,13 +1,20 @@
 from unittest.mock import ANY
 
+import numpy as np
 import pytest
 from ophyd_async.core import (
     DeviceCollector,
     assert_configuration,
     assert_reading,
+    set_mock_value,
 )
 
-from dodal.devices.undulator import Undulator, UndulatorGapAccess
+from dodal.devices.undulator import (
+    Undulator,
+    UndulatorGapAccess,
+    _get_closest_gap_for_energy,
+)
+from dodal.devices.undulator_dcm import AccessError
 
 ID_GAP_LOOKUP_TABLE_PATH: str = (
     "./tests/devices/unit_tests/test_beamline_undulator_to_gap_lookup_table.txt"
@@ -105,3 +112,22 @@ async def test_length_not_propagated_if_not_supplied():
         )
     assert undulator.length is None
     assert "undulator-length" not in (await undulator.read_configuration())
+
+
+@pytest.mark.parametrize(
+    "dcm_energy, expected_output", [(5730, 5.4606), (7200, 6.045), (9000, 6.404)]
+)
+def test_correct_closest_distance_to_energy_from_table(dcm_energy, expected_output):
+    energy_to_distance_table = np.array([[5700, 5.4606], [7000, 6.045], [9700, 6.404]])
+    assert (
+        _get_closest_gap_for_energy(dcm_energy, energy_to_distance_table)
+        == expected_output
+    )
+
+
+async def test_when_gap_access_is_disabled_set_energy_then_error_is_raised(
+    undulator,
+):
+    set_mock_value(undulator.gap_access, UndulatorGapAccess.DISABLED)
+    with pytest.raises(AccessError):
+        await undulator.set(5)
